jpmorgan-quant-simulation/requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.1.0
matplotlib>=3.6.0
python-dateutil>=2.8.2
